优化tag方式，优化医生药品，优化药品基础数据（中西药管理，单次剂量和单位的处理），删除无用代码，增加从平台获取药品数据的可能

+++ hospitalService/core/src/main/java/com/yyg/eprescription/entity/PrescriptionDrugs.java
@@ -37,16 +37,16 @@ public class PrescriptionDrugs implements Serializable{
 	private Long prescriptionid;//处方编号
 	
 	@Column(name = "drugid")
-	@ColumnType(jdbcType = JdbcType.BIGINT)
+	@ColumnType(jdbcType = JdbcType.INTEGER)
 	@ApiModelProperty(value = "药品编号")
-	private Long drugid;//处方编号
+	private Integer drugid;//处方编号
 	
 	
-	public Long getDrugid() {
+	public Integer getDrugid() {
 		return drugid;
 	}
 
-	public void setDrugid(Long drugid) {
+	public void setDrugid(Integer drugid) {
 		this.drugid = drugid;
 	}
